@@ -11,20 +11,20 @@
   "license": "ISC",
   "dependencies": {
     "axios": "^0.16.2",
-    "react": "^15.5.0",
+    "react": "^15.6.1",
     "react-bootstrap": "^0.31.0",
     "react-d3-basic": "^1.6.11",
-    "react-dom": "^15.5.0",
+    "react-dom": "^15.6.1",
     "react-router": "^4.1.1"
   },
   "devDependencies": {
-    "babel-core": "^6.24.0",
-    "babel-loader": "^6.4.1",
-    "babel-preset-react": "^6.23.0",
-    "eslint": "^4.0.0",
-    "eslint-plugin-react": "^7.1.0",
+    "babel": "^6.23.0",
+    "babel-core": "^6.25.0",
+    "babel-loader": "^7.1.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
     "html-webpack-plugin": "^2.28.0",
-    "webpack": "^2.3.1",
-    "webpack-dev-server": "^2.4.2"
+    "webpack": "^3.0.0",
+    "webpack-dev-server": "^2.5.0"
   }
 }
