@@ -1,32 +1,30 @@
 var React = require('react');
 var PartyInstance = require('./PartyInstance');
 var Link = require('react-router-dom').Link;
+var partyFile = require('../data/parties.json');
 
-var Party = React.createClass({
-	render: function () {
-		parties = [
-			{key:"national", name: "National", colour:"blue"},
-			{key:"labour", name: "Labour", colour:"red"},
-			{key:"greens", name: "Greens", colour:"green"} 
-		]
+class Party extends React.Component {
+
+	constructor(props){
+		super(props);
+	}
+
+	render() {
+		var parties = partyFile;
+
 		return (
-			<div style={{align:'justify'}}>
-				{parties.map(function(party){
-				return( 
-				<Link key={party.key} to={"/party/"+party.key}>
-					<PartyInstance name={party.name} colour={party.colour} />
-				</Link>
-				)})}
-				<p>
-					Do some database magic to fetch list of parties
-
-					Clicking on one opens info page with info sheets -- list MPs, current seats, etc. 
-
-					National, Labour, etc
-				</p>
+			<div>
+				{Object.keys(parties).map(function(party){
+					return( 
+						<Link key={party} to={"/party/"+party}>
+							<PartyInstance name={parties[party].name} image={parties[party].image}/>
+						</Link>
+					)
+				})
+				}
 			</div>
-			)
+		);
+	}
 }
-})
 
 module.exports = Party;
