@@ -1,14 +1,16 @@
 var React = require('react');
+var partyFile = require('../data/parties.json');
+var pollingFile = require('../data/latestpoll.json');
+var SeatInputs = require('./SeatInputs');
+var SeatDiagram = require('./SeatDiagram');
 
 var Polling = React.createClass({
 	render: function () {
 		return (
 			<div>
-				<p>
-					Poooollls.
-					Find a nice graphing library.
-					Do more DB magic.
-				</p>
+				<SeatInputs />
+
+				<SeatDiagram parties={partyFile} polling={pollingFile.pollData} />
 			</div>
 			)
 	}
